refactor(nostr): put a NostrRelayBackend seam under the relay layer
Phase 5 of docs/specs/nostr-sdk-migration.md. Pure refactor: the transport
becomes swappable, and nothing about the app's behavior changes.

The interface is deliberately narrow — connect, subscribe, publish to one
relay, report what that relay said. Everything that makes publishing
*trustworthy* stayed above it in NostrService: the outbox, per-relay
convergence and its ack bookkeeping, the resend sweep, supersession,
monotonic created_at, the addressable cache, NIP-40 filtering.

That split is the point of the phase. Each of those behaviors was bought with
a production bug (#77, #78), and none of them is a property of a transport.
Pushed down into the backend, they would have to be re-implemented and
re-proven against nostr-sdk's relay pool in Phase 6 — and the bug that started
this whole thread, a relay quietly serving a stale scoreboard, would have had
a fresh place to hide.

What the transport does own is what is genuinely about sockets: the
OK-confirmed publish, and treating a silent relay as dead rather than slow.

RelayConnection moved verbatim into DartRelayBackend, subtleties intact, and
its tests moved with it into relay/dart_relay_backend_test.dart — where they
read as what they are: transport regressions, not service tests.

NostrService no longer imports web_socket_channel; it can no longer form an
opinion about sockets at all. 165 tests green, analyzer unchanged.
